ci: remove tensorflow from testoptim
